+++ lib/src/validators/phone.dart
@@ -126,11 +126,13 @@ class Phone {
   ///```
   /// ### Parameters:
   /// - [mobileNumber]: The mobile number to identify the operator for. The number should be in a valid format.
+  /// - [size]: The size of the icon, default is 24.
   ///
   /// ### Returns:
   /// - The icon of the mobile operator as a widget if the mobile number is valid and matches an operator code.
   /// - `null` if the number is invalid or does not match any operator.
-  static Widget? getMobileOperatorIcon(String mobileNumber) {
+  static Widget? getMobileOperatorIcon(String mobileNumber,
+      {double size = 24}) {
     if (isMobileNumberValid(mobileNumber)) {
       mobileNumber = mobileNumber.replaceAll(RegExp(r'\s+'), '');
       if (mobileNumber.startsWith('+98')) {
@@ -141,7 +143,10 @@ class Phone {
       for (var operator in _mobileOperators) {
         for (var code in operator.numberCode) {
           if (mobileNumber.startsWith(code)) {
-            return MobileOperatorIcon(operator: operator);
+            return MobileOperatorIcon(
+              operator: operator,
+              size: size,
+            );
           }
         }
       }